django-storage-upload-image/requirements.txt
django==5.2.4
djangorestframework==3.16.0
django-storages[s3,google]==1.14.6
pillow==10.2.0
django-cors-headers==4.3.1
